Rename enums.py to utils.py
